feat(app): implement interactive walkthrough graph view
Extracted graph layout logic from the objective curation view into a reusable `useGraphLayout` hook. Created a new `WalkthroughGraph` component for visualizing a guide's prerequisite DAG.

Added the `WalkthroughGraph` component to the walkthrough route so users can interactively browse dependencies and navigate directly to prerequisites.
