+++ app/src/components/graph-view/GuideNode.tsx
